news_recommender/requirements.txt
Flask
Flask_SQLAlchemy
Flask_CORS
python-dotenv
psycopg2-binary
PyJWT
scikit-learn